+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 + [x] Utils: removed function `isLocalhost`, if you need, use `$.localhost` or `$.isLocalhost([host])`.
 + [x] Utils: removed function `isArray`.
 + [x] Utils: removed function `isTouchDevice`. if you need, use `$.touchable`.
++ [x] Tag input: improve method `val()`. Now you can use for new value(s) `array` or `simple value`.
++ [x] Tag input: added method `append(val)`. For append tag(s) to existing value.
 
 ### 4.4.0
 #### M4Q